Predict `C_PasstimeGun::m_eThrowState` on the client
This value is truly predicted across the server and client (much of the
code relying on this value is executed on both side). If the client's
value were to become desynced, there would be nothing to enforce it's
value unless it changed on the server again.

Add it to the prediction table, so that mispredictions are caught
proper.