Fix lock-on issues by making `tf_passtime_mode_lock_eye_to_eye` LOS only
Previously, we overrode the entire target position when eye-to-eye
passing is enabled. This would cause issues when up-close to a player,
as the target position you had to look at was now the player's eyes.

Resolve this by only using the eye-to-eye lock-on position for LOS
checks -- the other positions are left as is upstream.

While we are here, move and rename our convar to
`p4ss_lock_eye_to_eye_los`.